chore: consolidate validator functions into fewer files
